// client/src/actions/actionType.ts
export enum ActionType {
	SET_ALERT = 'SET_ALERT',
	REMOVE_ALERTS = 'REMOVE_ALERTS',
	REGISTER_SUCCESS = 'REGISTER_SUCCESS',
	REGISTER_FAIL = 'REGISTER_FAIL',
	LOAD_USER = 'LOAD_USER',
	AUTH_FAILED = 'AUTH_FAILED',
	LOGIN_SUCCESS = 'LOGIN_SUCCESS',
	LOGIN_FAIL = 'LOGIN_FAIL',
	LOGOUT_SUCCESS = 'LOGOUT_SUCCESS',
	GET_PROFILE = 'GET_PROFILE',
	GET_PROFILES = 'GET_PROFILES',
	GET_REPOS = 'GET_REPOS',
	PROFILE_FAIL = 'PROFILE_FAIL',
	PROFILES_FAIL = 'PROFILES_FAIL',
	REPOS_FAIL = 'REPOS_FAIL',
	CLEAR_PROFILE = 'CLEAR_PROFILE',
	DELETE_PROFILE = 'DELETE_PROFILE',
	GET_POSTS = 'GET_POSTS',
	POSTS_FAIL = 'POSTS_FAIL',
	POST_FAIL = 'POST_FAIL',
	GET_POST = 'GET_POST',
	ADD_POST = 'ADD_POST',
	ADD_COMMENT = 'ADD_COMMENT',
	REMOVE_POST = 'REMOVE_POST',
	LIKE_POST = 'LIKE_POST',
}